Fix circular dependencies 😎
