+++ glean/src/core/metrics/types/uuid.ts
@@ -7,7 +7,7 @@ import { MetricType } from "../index.js";
 import { generateUUIDv4, isString } from "../../utils.js";
 import { Context } from "../../context.js";
 import { Metric } from "../metric.js";
-import { ErrorType, recordError, testGetNumRecordedErrors } from "../../error_recording.js";
+import { ErrorType } from "../../error/error_type.js";
 
 // Loose UUID regex for checking if a string has a UUID _shape_. Does not contain version checks.
 //
@@ -68,7 +68,7 @@ class UUIDMetricType extends MetricType {
     try {
       metric = new UUIDMetric(value);
     } catch {
-      await recordError(
+      await Context.errorManager.record(
         instance,
         ErrorType.InvalidValue,
         `"${value}" is not a valid UUID.`
@@ -126,19 +126,6 @@ class UUIDMetricType extends MetricType {
       metric = await Context.metricsDatabase.getMetric<string>(ping, this);
     });
     return metric;
-  }
-
-  /**
-   * Returns the number of errors recorded for the given metric.
-   *
-   * @param errorType The type of the error recorded.
-   * @param ping represents the name of the ping to retrieve the metric for.
-   *        Defaults to the first value in `sendInPings`.
-   *
-   * @returns the number of errors recorded for the metric.
-   */
-  async testGetNumRecordedErrors(errorType: string, ping: string = this.sendInPings[0]): Promise<number> {
-    return testGetNumRecordedErrors(this, errorType as ErrorType, ping);
   }
 }
