+++ glean/tests/core/error_recording.spec.ts
@@ -4,7 +4,8 @@
 
 import assert from "assert";
 
-import { ErrorType, recordError, testGetNumRecordedErrors } from "../../src/core/error_recording";
+import { ErrorType } from "../../src/core/error/error_type";
+import { Context } from "../../src/core/context";
 import Glean from "../../src/core/glean";
 import { Lifetime } from "../../src/core/metrics/lifetime";
 import LabeledMetricType from "../../src/core/metrics/types/labeled";
@@ -26,11 +27,11 @@ describe("error_recording", function() {
       disabled: false
     });
 
-    await recordError(metric, ErrorType.InvalidValue, "Invalid value");
-    await recordError(metric, ErrorType.InvalidLabel, "Invalid label", 10);
+    await Context.errorManager.record(metric, ErrorType.InvalidValue, "Invalid value");
+    await Context.errorManager.record(metric, ErrorType.InvalidLabel, "Invalid label", 10);
 
-    assert.strictEqual(await testGetNumRecordedErrors(metric, ErrorType.InvalidValue), 1);
-    assert.strictEqual(await testGetNumRecordedErrors(metric, ErrorType.InvalidLabel), 10);
+    assert.strictEqual(await Context.errorManager.testGetNumRecordedErrors(metric, ErrorType.InvalidValue), 1);
+    assert.strictEqual(await Context.errorManager.testGetNumRecordedErrors(metric, ErrorType.InvalidLabel), 10);
   });
 
   it("strips label when recording error to metric that contains label in it's name", async function () {
@@ -48,10 +49,10 @@ describe("error_recording", function() {
 
     // When the metric is created from a static label, it will contain the label in it's name,
     // otherwise the label stays underd the `dynamicLabel` property.
-    await recordError(metric.oneLabel, ErrorType.InvalidValue, "Invalid value");
-    await recordError(metric.anotherLabel, ErrorType.InvalidLabel, "Invalid label", 10);
+    await Context.errorManager.record(metric.oneLabel, ErrorType.InvalidValue, "Invalid value");
+    await Context.errorManager.record(metric.anotherLabel, ErrorType.InvalidLabel, "Invalid label", 10);
 
-    assert.strictEqual(await testGetNumRecordedErrors(metric.dynamicLabel, ErrorType.InvalidValue), 1);
-    assert.strictEqual(await testGetNumRecordedErrors(metric.dynamicLabel, ErrorType.InvalidLabel), 10);
+    assert.strictEqual(await Context.errorManager.testGetNumRecordedErrors(metric.dynamicLabel, ErrorType.InvalidValue), 1);
+    assert.strictEqual(await Context.errorManager.testGetNumRecordedErrors(metric.dynamicLabel, ErrorType.InvalidLabel), 10);
   });
 });
