+++ glean/tests/core/metrics/counter.spec.ts
@@ -4,7 +4,7 @@
 
 import assert from "assert";
 import { Context } from "../../../src/core/context";
-import { ErrorType } from "../../../src/core/error_recording";
+import { ErrorType } from "../../../src/core/error/error_type";
 
 import Glean from "../../../src/core/glean";
 import { Lifetime } from "../../../src/core/metrics/lifetime";
