+++ glean/tests/core/metrics/labeled.spec.ts
@@ -6,7 +6,7 @@ import assert from "assert";
 import sinon from "sinon";
 
 import { Context } from "../../../src/core/context";
-import { ErrorType } from "../../../src/core/error_recording";
+import { ErrorType } from "../../../src/core/error/error_type";
 import Glean from "../../../src/core/glean";
 import { Lifetime } from "../../../src/core/metrics/lifetime";
 import BooleanMetricType from "../../../src/core/metrics/types/boolean";
