+++ glean/tests/core/metrics/string.spec.ts
@@ -9,7 +9,7 @@ import StringMetricType from "../../../src/core/metrics/types/string";
 import { Lifetime } from "../../../src/core/metrics/lifetime";
 import { MAX_LENGTH_VALUE } from "../../../src/core/metrics/types/string";
 import { Context } from "../../../src/core/context";
-import { ErrorType } from "../../../src/core/error_recording";
+import { ErrorType } from "../../../src/core/error/error_type";
 
 describe("StringMetric", function() {
   const testAppId = `gleanjs.test.${this.title}`;
